+++ src/utils/color/Palette.ts
@@ -1,3 +1,7 @@
+/**************************************
+ * by: Joaopssx - xdcanvas
+ * https://github.com/joaopssx/xdcanvas
+ **************************************/
 import { Conversions } from './Conversions';
 
 export class Palette {
@@ -32,3 +36,4 @@ export class Palette {
     return colors;
   }
 }
+