+++ src/utils/color/types.ts
@@ -1,4 +1,9 @@
+/**************************************
+ * by: Joaopssx - xdcanvas
+ * https://github.com/joaopssx/xdcanvas
+ **************************************/
 export interface RGB { r: number; g: number; b: number; a?: number; }
 export interface HSL { h: number; s: number; l: number; }
 export interface HSV { h: number; s: number; v: number; }
 export interface CMYK { c: number; m: number; y: number; k: number; }
+